+++ docs/changelog.md
@@ -224,10 +224,17 @@ Many feature gates are removed in order to make most features make always availa
 - move to [base]:
   - traits: `NumToStr`, [`UnicodeScalar`].
   - types: `AsciiDigits`, `ByteSearch`, `GraphemeNonul`, `GraphemeU*`, `Str`, `StringNonul`, `StringU*`, `char7`, `char8`, `char16`.
+
+### ascii
 - rename `Ascii` to `AsciiDigits`.
   - rename method: `calc_digit` to `digit_at_power`.
-- new `char7` methods: `to_byte`, `to_str`.
 - make `ASCII_TABLE` public.
+
+### char
+- new `char7` methods: `to_byte`, `to_str`.
+- update `Char`:
+  - remove deprecated methods: `len_to_utf8`, `utf8_?bytes_len`.
+  - return lenghts as usize.
 
 ### fmt
 - move to [base]: